% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldenCheetah_AWSS3.R
\name{between.gc_opendata_athletes}
\alias{between.gc_opendata_athletes}
\title{Return athlete IDs that have been modified between two dates}
\usage{
between.gc_opendata_athletes(
  object = NULL,
  from = -Inf,
  to = +Inf,
  details = TRUE
)
}
\arguments{
\item{object}{an \code{object} of class \code{gc_opendata_athletes}. If \code{NULL} (default) then the output of \code{get_athletes(TRUE)} is used}

\item{from}{a character string of the form "YYYY-MM-DD" specifying the from date. If \code{-Inf} (default) then the minimum date from \code{object$LastModified}}

\item{to}{a character string of the form "YYYY-MM-DD" specifying the to date. If \code{Inf} (default) then the maximum date from \code{object$LastModified}}

\item{details}{if \code{FALSE} a vector with the available athlete IDs is returned. If \code{TRUE}, then a data frame with athlete IDs and accompnying details is returned. See Details.}
}
\description{
Return athlete IDs that have been modified between two dates
}
