% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{between.GCOD_df}
\alias{between.GCOD_df}
\title{Return athlete IDs that have been modified between two dates.}
\usage{
between.GCOD_df(object = NULL, from = -Inf, to = +Inf)
}
\arguments{
\item{object}{an \code{object} of class \code{GCOD_df}, as produced by \code{\link{get_athlete_ids}}. If \code{NULL} (default) then the output of a call to \code{\link[=get_athletes]{get_athletes()}} is used.}

\item{from}{a character string of the form "YYYY-MM-DD" specifying the from date. If \code{-Inf} (default) then the minimum date from \code{object$LastModified}}

\item{to}{a character string of the form "YYYY-MM-DD" specifying the to date. If \code{Inf} (default) then the maximum date from \code{object$LastModified}}
}
\description{
Return athlete IDs that have been modified between two dates.
}
