% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcod_db.R
\name{check_local_dir.gcod_db}
\alias{check_local_dir.gcod_db}
\title{Checks whether the files in the remote paths of a \code{gcod_db} exist in `local_dir#'}
\usage{
check_local_dir.gcod_db(object, local_dir)
}
\arguments{
\item{object}{asd}

\item{local_dir}{the directory to check zip files for the selected athlete IDs.}
}
\description{
Checks whether the files in the remote paths of a \code{gcod_db} exist in `local_dir#'
}
\details{
Will update local_db accordingly
Ignores the contents of local_db, and rebuilds it according to the contents of local_dir
}
