% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_athlete_data.R
\name{download_workouts}
\alias{download_workouts}
\title{Download and, optionally, extract the archive with the workouts for a particular athlete ID}
\usage{
download_workouts(
  athlete_id,
  dir = tempdir(),
  pattern = NULL,
  extract = FALSE,
  mirror = "S3",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{athlete_id}{a character string with the athlete ID or the first few characters of it, or an object of class \code{GCOD_df} as constructed by \code{get_athlete_ids()}.}

\item{dir}{the directory to download the zip files for the selected athleted IDs.}

\item{pattern}{character string containing a regular expression to be matched with the athlete IDs in \code{athlete_id}. Only applicable if \code{athlete_id} is an object of class \code{GCOD_df}. Default is \code{NULL}, which selects all IDs in \code{athlete_id}.}

\item{extract}{logical determining whether the workout files in the downloaded archives should be extracted. Default is \code{FALSE}. If \code{TRUE}, then the archives are extractred in sub-directories unded \code{dir}. The sub-directories are named according to the athlete ID.}

\item{mirror}{either \code{"S3"} or \code{"OSF"}, indicating the GoldeCheetah OpenData mirror to use. Default and recommended is "S3". See Details.}

\item{verbose}{logical determining whether progress information should be printed. Default is \code{FALSE}.}

\item{...}{extra arguments to be passed to \code{aws.s3::save_object()}.}
}
\description{
Download and, optionally, extract the archive with the workouts for a particular athlete ID
}
