% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_management.R
\name{exist_in.gcod_db}
\alias{exist_in.gcod_db}
\title{Checks whether the files in the remote paths of a \code{gcod_db} exist in `local_dir'}
\usage{
exist_in.gcod_db(object, local_dir)
}
\arguments{
\item{object}{asd}

\item{local_dir}{the directory to check zip files for the selected athlete IDs.}
}
\description{
Checks whether the files in the remote paths of a \code{gcod_db} exist in `local_dir'
}
