% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldenCheetahOpenData-package.R,
%   R/extract_workouts.R
\name{extract_workouts}
\alias{extract_workouts}
\alias{extract_workouts.gcod_db}
\title{Extract workouts from the workout archives downloaded using \code{\link[=download_workouts]{download_workouts()}}.}
\usage{
extract_workouts(object, ...)

\method{extract_workouts}{gcod_db}(
  object,
  verbose = FALSE,
  clean_up = TRUE,
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{object}{an object of class \code{gcod_db} as produced from \code{\link[=download_workouts]{download_workouts()}}.}

\item{...}{currently not used.}

\item{verbose}{logical determining whether progress information should be printed. Default is \code{FALSE}.}

\item{clean_up}{logical determining whether the workout directories should be deleted before extraction, if they already exist. Default is \code{FALSE}.}

\item{overwrite}{logical determining whether the workout directories should be overwritten, if they already exist. Default is \code{TRUE}.}
}
\value{
An object of class \code{gcod_db} which is the same as \code{object} except
that the elements of \code{local_perspective(object)$extracted} are set to \code{TRUE} or
\code{FALSE} depending on whether the corresponding workout archives
were extracted successfully or not.
}
\description{
Extract workouts from the workout archives downloaded using \code{\link[=download_workouts]{download_workouts()}}.
}
\details{
Athlete IDs are infered from \code{local_path(object)}.

' The workouts are extracted and put in sub-directories in the same
directory as the workout archives (i.e. where the files
\code{local_path(object)} are). These sub-directories have exactly the
same name as the archives (excluding the file extension).
}
\seealso{
\code{\link[=download_workouts]{download_workouts()}}
}
