% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_workouts.R
\name{extract_workouts.gcod_db}
\alias{extract_workouts.gcod_db}
\title{Extract workouts from the archives downloaded using \code{\link[=download_workouts]{download_workouts()}}.}
\usage{
\method{extract_workouts}{gcod_db}(object, verbose = FALSE, clean_up = FALSE, overwrite = TRUE)
}
\arguments{
\item{object}{an object of class \code{gcod_db} as produced from \code{\link[=download_workouts]{download_workouts()}}.}

\item{verbose}{logical determining whether progress information should be printed. Default is \code{FALSE}.}

\item{clean_up}{logical determining whether the workout directories should be deleted before extraction, if they already exist. Default is \code{FALSE}.}

\item{overwrite}{logical determining whether the workout directories should be overwritten, if they already exist. Default is \code{TRUE}.}
}
\value{
An object of class \code{gcod_db} which is the same as \code{object} except that \code{object$extracted = TRUE}.

The workouts are extracted and put in sub-directories in the same directory as the original archive (i.e. where the files \code{object$path} are). These sub-directories have exactly the same name as the archives (excluding the file extension).
}
\description{
Extract workouts from the archives downloaded using \code{\link[=download_workouts]{download_workouts()}}.
}
\details{
Athlete IDs are infered from \code{object$path}.
}
\seealso{
\code{\link[=download_workouts]{download_workouts()}}
}
