% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcod_db.R
\name{gcod_db_extractors}
\alias{gcod_db_extractors}
\alias{min_size}
\alias{max_size}
\alias{mean_size}
\alias{total_size}
\alias{n_ids}
\alias{n_ids.gcod_db}
\alias{min_size.gcod_db}
\alias{max_size.gcod_db}
\alias{total_size.gcod_db}
\alias{mean_size.gcod_db}
\alias{local_path}
\title{Extract information from an object of class \code{gcod_db}, as produced by \code{\link[=get_athlete_ids]{get_athlete_ids()}}.}
\usage{
\method{n_ids}{gcod_db}(object, db = "remote")

\method{min_size}{gcod_db}(object, db = "remote")

\method{max_size}{gcod_db}(object, db = "remote")

\method{total_size}{gcod_db}(object, db = "remote")

\method{total_size}{gcod_db}(object, db = "remote")

\method{mean_size}{gcod_db}(object, db = "remote")

local_path(object)
}
\arguments{
\item{object}{an object of class \code{gcod_db}, as produced by \code{\link[=get_athlete_ids]{get_athlete_ids()}}.}
}
\description{
Extract information from an object of class \code{gcod_db}, as produced by \code{\link[=get_athlete_ids]{get_athlete_ids()}}.
}
\details{
File sizes are reported in bytes but inherit from class \code{object_size}. So, \code{format.object_size()} can be used for pretty units, etc.
}
\seealso{
\code{\link[=get_athlete_ids]{get_athlete_ids()}}
}
