% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GCOD_df_info}
\alias{GCOD_df_info}
\alias{min_size}
\alias{max_size}
\alias{mean_size}
\alias{total_size}
\alias{n_ids}
\alias{n_ids.GCOD_df}
\alias{min_size.GCOD_df}
\alias{max_size.GCOD_df}
\alias{total_size.GCOD_df}
\alias{mean_size.GCOD_df}
\title{Extract information from an object of class \code{GCOD_df}, as produced by \code{\link[=get_athlete_ids]{get_athlete_ids()}}.}
\usage{
\method{n_ids}{GCOD_df}(object)

\method{min_size}{GCOD_df}(object)

\method{max_size}{GCOD_df}(object)

\method{total_size}{GCOD_df}(object)

\method{mean_size}{GCOD_df}(object)
}
\arguments{
\item{object}{an object of class \code{GCOD_df}, as produced by \code{\link[=get_athlete_ids]{get_athlete_ids()}}.}
}
\description{
Extract information from an object of class \code{GCOD_df}, as produced by \code{\link[=get_athlete_ids]{get_athlete_ids()}}.
}
\details{
File sizes are reported in bytes but inherit from class \code{object_size}. So \code{format.object_size()} can be used for pretty units, etc.
}
\seealso{
\code{\link[=get_athlete_ids]{get_athlete_ids()}}
}
