% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athlete_ids.R
\name{print.GCOD_df}
\alias{print.GCOD_df}
\title{Print method for objects of class \code{GCOD_df}, as produced by \code{\link[=get_athlete_ids]{get_athlete_ids()}}.}
\usage{
\method{print}{GCOD_df}(object, txtplot = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{GCOD_df}.}

\item{txtplot}{logical indicating whether or not a text barplot should be printed of the percent of athlete ID records modified per year quarter. Default is \code{FALSE}.}

\item{...}{currently not used.}
}
\description{
Print method for objects of class \code{GCOD_df}, as produced by \code{\link[=get_athlete_ids]{get_athlete_ids()}}.
}
