% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_workouts.R
\name{read_workouts.GCOD_files}
\alias{read_workouts.GCOD_files}
\title{Read the workout \code{.csv} files that have been extracted using \code{\link[=extract_workouts]{extract_workouts()}} and organized them into lists of \code{\link[trackeR:trackeRdata]{trackeR::trackeRdata}} objects for further processing into the \code{trackeR} R package.}
\usage{
\method{read_workouts}{GCOD_files}(object, verbose = FALSE, delete = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{GCOD_files} as produced from \code{\link[=download_workouts]{download_workouts()}} or \code{\link[=extract_workouts]{extract_workouts()}}.}

\item{verbose}{logical determining whether progress information should be printed. Default is \code{FALSE}.}

\item{...}{other arguments to be passed to \code{\link[=extract_workouts]{extract_workouts()}}}
}
\value{
A list of \code{\link[trackeR:trackeRdata]{trackeR::trackeRdata}} objects.
}
\description{
Read the workout \code{.csv} files that have been extracted using \code{\link[=extract_workouts]{extract_workouts()}} and organized them into lists of \code{\link[trackeR:trackeRdata]{trackeR::trackeRdata}} objects for further processing into the \code{trackeR} R package.
}
\details{
If \code{object$extracted} is \code{FALSE}, then the workout files are extracted automatically using \code{\link[=extract_workouts]{extract_workouts()}}.

It is assumed that the filename for each workout corresponds to the timestamp where the first observation is made for the session. Timestamps are in UTC.

If the number of workout files in the archive of a particular athlete ID does not match the number of workouts recorded in the json files within that archive, then the workout files for the ID are not read.
}
\references{
Frick, H., Kosmidis, I. (2017). trackeR: Infrastructure for Running
and Cycling Data from GPS-Enabled Tracking Devices in R. \emph{Journal
of Statistical Software}, \strong{82}(7),
1--29. \href{https://doi.org/10.18637/jss.v082.i07}{doi:10.18637/jss.v082.i07}
}
\seealso{
\code{\link[trackeR:trackeRdata]{trackeR::trackeRdata()}}
}
