% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_management.R
\name{rebuild_gcod_db.character}
\alias{rebuild_gcod_db.character}
\title{Attempts to rebuild a \code{gcod_db} from the contents of a local directory}
\usage{
\method{rebuild_gcod_db}{character}(object, mirror = "S3")
}
\arguments{
\item{object}{a character string giving the path to the directory to use for extracting athlete IDs.}
}
\description{
Attempts to rebuild a \code{gcod_db} from the contents of a local directory
}
