% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_management.R
\name{c.gcod_db}
\alias{c.gcod_db}
\title{Concatanate gcod_db objects}
\usage{
\method{c}{gcod_db}(..., db = "remote")
}
\description{
Concatanate gcod_db objects
}
\details{
If \code{db = "remote"}, then \code{...$remote} are concatanated and
\code{...$local_db} from the first object is used. If \code{db = "local"},
then \code{...$local} are concatanated and \code{...$remote_db} from the
first object is used. If \code{db = "both"} then both  from \code{...$remote} and
\code{...$local_db} are concatanated.

Records with duplicated athlete IDs in the objects being
concatenated are replaced, in the concatanated object, by a single
record with the most recently modified entry.
}
