% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldenCheetahOpenData-package.R, R/db_methods.R
\name{read_workouts}
\alias{read_workouts}
\alias{n_ids}
\alias{min_size}
\alias{max_size}
\alias{total_size}
\alias{mean_size}
\alias{local_path}
\alias{remote}
\alias{local}
\alias{athlete_id}
\alias{rebuild_gcod_db}
\alias{gcod_db_extractors}
\alias{n_ids.gcod_db}
\alias{min_size.gcod_db}
\alias{max_size.gcod_db}
\alias{total_size.gcod_db}
\alias{mean_size.gcod_db}
\alias{local_path.gcod_db}
\alias{remote.gcod_db}
\alias{local.gcod_db}
\alias{athlete_id.gcod_db}
\title{Extract information from an object of class \code{gcod_db}, as produced
by \code{\link[=get_athlete_ids]{get_athlete_ids()}}.}
\usage{
read_workouts(object, ...)

n_ids(object, ...)

min_size(object, ...)

max_size(object, ...)

total_size(object, ...)

mean_size(object, ...)

local_path(object, ...)

remote(object, ...)

local(object, ...)

athlete_id(object, ...)

rebuild_gcod_db(object, ...)

\method{n_ids}{gcod_db}(object, perspective = "remote", ...)

\method{min_size}{gcod_db}(object, perspective = "remote", ...)

\method{max_size}{gcod_db}(object, perspective = "remote", ...)

\method{total_size}{gcod_db}(object, perspective = "remote", ...)

\method{mean_size}{gcod_db}(object, perspective = "remote", ...)

\method{local_path}{gcod_db}(object, ...)

\method{remote}{gcod_db}(object, ...)

\method{local}{gcod_db}(object, ...)

\method{athlete_id}{gcod_db}(object, perspective = "remote", ...)
}
\arguments{
\item{object}{an object of class \code{gcod_db}, as produced by
\code{\link[=get_athlete_ids]{get_athlete_ids()}}.}

\item{...}{currently not used.}

\item{perspective}{either \code{"remote"} (default) or \code{"local"}, for
the perspective to use for the extractor function.}
}
\description{
Extract information from an object of class \code{gcod_db}, as produced
by \code{\link[=get_athlete_ids]{get_athlete_ids()}}.
}
\details{
File sizes from the \code{min_size()}, \code{max_size()}, \code{total_size()},
\code{mean_size()} extractors are reported in bytes but inherit from
class \code{object_size}. So, \code{format.object_size()} can be used
directly for pretty units, etc.

\code{local_path()} extracts the local paths from the \code{gcod_db} file.

\code{remote(object)} and \code{local(object)} are equivalent to
\code{object$remote_db} and \code{object$local_db}, respectively.

\code{n_ids()} returns the number of athlete ids in the specified
perspective, and \code{athlete_id()} the athlete IDs themselves.
}
\seealso{
\code{\link[=get_athlete_ids]{get_athlete_ids()}}
}
