% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_methods.R
\name{gcod_db_extractors}
\alias{gcod_db_extractors}
\alias{min_size}
\alias{max_size}
\alias{mean_size}
\alias{total_size}
\alias{n_ids}
\alias{n_ids.gcod_db}
\alias{min_size.gcod_db}
\alias{max_size.gcod_db}
\alias{total_size.gcod_db}
\alias{mean_size.gcod_db}
\alias{local_path.gcod_db}
\alias{remote.gcod_db}
\alias{local.gcod_db}
\alias{athlete_id.gcod_db}
\title{Extract information from an object of class \code{gcod_db}, as produced by \code{\link[=get_athlete_ids]{get_athlete_ids()}}.}
\usage{
\method{n_ids}{gcod_db}(object, perspective = "remote")

\method{min_size}{gcod_db}(object, perspective = "remote")

\method{max_size}{gcod_db}(object, perspective = "remote")

\method{total_size}{gcod_db}(object, perspective = "remote")

\method{total_size}{gcod_db}(object, perspective = "remote")

\method{mean_size}{gcod_db}(object, perspective = "remote")

\method{local_path}{gcod_db}(object, ...)

\method{remote}{gcod_db}(object, ...)

\method{local}{gcod_db}(object, ...)

\method{athlete_id}{gcod_db}(object, perspective = "remote")
}
\arguments{
\item{object}{an object of class \code{gcod_db}, as produced by \code{\link[=get_athlete_ids]{get_athlete_ids()}}.}

\item{perspective}{either \code{"remote"} (default) or \code{"local"}, for the perspective to use for the extractor function.}
}
\description{
Extract information from an object of class \code{gcod_db}, as produced by \code{\link[=get_athlete_ids]{get_athlete_ids()}}.
}
\details{
File sizes are reported in bytes but inherit from class \code{object_size}. So, \code{format.object_size()} can be used for pretty units, etc.

\code{local_path()} extracts the local paths from the \code{gcod_db} file. \code{remote(object)} and \code{local(object)} are equivalent to \code{object$remote_db} and \code{object$local_db}, respectively.
}
\seealso{
\code{\link[=get_athlete_ids]{get_athlete_ids()}}
}
