% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athlete_ids.R
\name{get_athlete_ids}
\alias{get_athlete_ids}
\title{Retrieve information for athlete IDs that are available in the GoldenCheetah OpenData project.}
\usage{
get_athlete_ids(n_ids = Inf, mirror = "S3", prefix = NULL, ...)
}
\arguments{
\item{n_ids}{integer indicating the maximum number of athlete IDs to return. Default is \code{Inf}, which will return all available athelte IDs.}

\item{mirror}{either \code{"S3"} or \code{"OSF"}, indicating which GoldeCheetah OpenData mirror should be used. Default and recommended is "S3". See Details.}

\item{prefix}{character string that limits the response to athlete IDs that begin with it. Default is \code{NULL}, which does not limit responses.}

\item{...}{further options to be passed to \verb{[aws.s3::get_bucket_df()]}.}
}
\value{
A \code{\link{list}} also inheriting from class \code{gcod_db} with components \code{remote_db} (of class \code{gcod_remote_db}) and \code{local_db} (of class \code{gcod_local_db}). \code{remote_db} and \code{local_db} are \code{\link{data.frame}}s, also inheriting from classes \code{gcod_remote_db} and \code{gcod_local_db}, respectively. \code{local_db} has variables "path", "last_modified", "size", "extracted", "downloaded", "athlete_id"m and \code{remote_db} has variables "key", "last_modified", "e_tag", "size", "owner_id", "owner_display_name", "storage_class", "bucket", and "athlete_id".
}
\description{
Retrieve information for athlete IDs that are available in the GoldenCheetah OpenData project.
}
\details{



}
\examples{

\donttest{
# Get details for all available athlete IDs
ids <- get_athlete_ids()
# Print basic info for the IDs retrieved
print(ids, txtplot = TRUE)

# Extract basic info for the IDs retrieved
n_ids(ids)
format(min_size(ids), unit = "auto")
format(max_size(ids), unit = "auto")
format(mean_size(ids), unit = "auto")
format(total_size(ids), unit = "auto")

# `ids` inherits from `data.frame` so we can easily do stuff like
# find which IDshave the minimum/maximum sizes
ids[which.min(ids$size), "athlete_id"]
ids[which.max(ids$size), "athlete_id"]
# find all IDs with the string "007"
subset(ids, grepl("007", ids$athlete_id))
# etc
}

}
\references{
Liversedge, M. (2020). GoldenCheetah OpenData Project. OSF. \url{https://doi.org/10.17605/OSF.IO/6HFPZ}
}
\seealso{
\code{\link[=print.gcod_db]{print.gcod_db()}} \code{\link[=min_size.gcod_db]{min_size.gcod_db()}} \code{\link[=max_size.gcod_db]{max_size.gcod_db()}} \code{\link[=total_size.gcod_db]{total_size.gcod_db()}} \code{\link[=mean_size.gcod_db]{mean_size.gcod_db()}} \code{\link[=n_ids.gcod_db]{n_ids.gcod_db()}}
}
