% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athlete_ids.R
\name{get_athlete_ids}
\alias{get_athlete_ids}
\title{List available athletes}
\usage{
get_athlete_ids(n_athletes = Inf, mirror = "S3", prefix = NULL, ...)
}
\arguments{
\item{n_athletes}{integer indicating the maximum number of athlete IDs to return. Default is \code{Inf}, which will return all available athelte IDs.}

\item{mirror}{either \code{"S3"} or \code{"OSF"}, indicating the GoldeCheetah OpenData mirror to use. Default and recommended is "S3". See Details.}

\item{prefix}{character string that limits the response to athlete IDs that begin with it. Default is \code{NULL}, which does not limit responses.}

\item{...}{further options to be passed to \verb{[aws.s3::get_bucket_df()]}.}
}
\description{
List available athletes
}
\details{
The \code{data.frame} that is returned if \code{details = TRUE} has
"Key", "LastModified", "ETag", "Size", "Size", "Owner_ID",
"Owner_DisplayName", "StorageClass", "Bucket", "athlete_id"

IK, 20200517: ADD DETAILS ON MIRROR

IK, 20200517: ADD EXAMPLES
}
