% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athlete_ids.R
\name{get_athlete_ids}
\alias{get_athlete_ids}
\title{Retrieve information for athlete IDs that are available in the GoldenCheetah OpenData project.}
\usage{
get_athlete_ids(n_ids = Inf, mirror = "S3", prefix = NULL, ...)
}
\arguments{
\item{n_ids}{integer indicating the maximum number of athlete IDs to return. Default is \code{Inf}, which will return all available athelte IDs.}

\item{mirror}{either \code{"S3"} or \code{"OSF"}, indicating the GoldeCheetah OpenData mirror to use. Default and recommended is "S3". See Details.}

\item{prefix}{character string that limits the response to athlete IDs that begin with it. Default is \code{NULL}, which does not limit responses.}

\item{...}{further options to be passed to \verb{[aws.s3::get_bucket_df()]}.}
}
\value{
A \code{\link{data.frame}} inheriting from class \code{GCOD_df} with variables
"key", "last_modified", "e_tag", "size", "owner_id",
"owner_display_name", "storage_class", "bucket", and "athlete_id".
}
\description{
Retrieve information for athlete IDs that are available in the GoldenCheetah OpenData project.
}
\details{

}
\examples{

\donttest{
# Get details for all available athlete IDs
ids <- get_athlete_ids()
# Print basic info for the IDs retrieved
print(ids, txtplot = TRUE)

# Extract basic info for the IDs retrieved
n_ids(ids)
format(min_size(ids), unit = "auto")
format(max_size(ids), unit = "auto")
format(mean_size(ids), unit = "auto")
format(total_size(ids), unit = "auto")

# `ids` inherits from `data.frame` so we can easily do stuff like
# find which IDshave the minimum/maximum sizes
ids[which.min(ids$size), "athlete_id"]
ids[which.max(ids$size), "athlete_id"]
# find all IDs with the string "007"
subset(ids, grepl("007", ids$athlete_id))
# etc
}

}
\references{
Liversedge, M. (2020). GoldenCheetah OpenData Project. OSF. \url{https://doi.org/10.17605/OSF.IO/6HFPZ}
}
\seealso{
\code{\link[=print.GCOD_df]{print.GCOD_df()}} \code{\link[=min_size.GCOD_df]{min_size.GCOD_df()}} \code{\link[=max_size.GCOD_df]{max_size.GCOD_df()}} \code{\link[=total_size.GCOD_df]{total_size.GCOD_df()}} \code{\link[=mean_size.GCOD_df]{mean_size.GCOD_df()}} \code{\link[=n_ids.GCOD_df]{n_ids.GCOD_df()}}
}
