% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldenCheetah_AWSS3.R
\name{get_athlete_ids}
\alias{get_athlete_ids}
\title{List available athletes}
\usage{
get_athlete_ids(details = TRUE, n_athletes = Inf, mirror = "S3")
}
\arguments{
\item{details}{if \code{FALSE} a vector with the available athlete IDs is returned. If \code{TRUE}, then a data frame with athlete IDs and accompnying details is returned. See Details.}
}
\description{
List available athletes
}
\details{
The \code{data.frame} that is returned if \code{details = TRUE} has
"Key", "LastModified", "ETag", "Size", "Size", "Owner_ID",
"Owner_DisplayName", "StorageClass", "Bucket", "athlete_id"
}
