% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldenCheetahOpenData-package.R
\docType{package}
\name{GoldenCheetahOpenData}
\alias{GoldenCheetahOpenData}
\title{GoldenCheetahOpenData: R API and local database management for the
GoldenCheetah OpenData Project}
\description{
GoldenCheetahOpenData provides methods for querying the GoldenCheetah
OpenData Project database (Liversedge, 2020), downloading workout
data from it and managing local workout databases. Methods are also
provided for the organization of the workout data into
'trackeRdata' objects for further data analysis and modelling in R
using the infrastructure provided by the \code{trackeR} R package
\url{https://CRAN.R-project.org/package=trackeR}.
}
\references{
Frick, H., Kosmidis, I. (2017). trackeR: Infrastructure for Running
and Cycling Data from GPS-Enabled Tracking Devices in R. \emph{Journal
of Statistical Software}, \strong{82}(7),
1--29. \href{https://doi.org/10.18637/jss.v082.i07}{doi:10.18637/jss.v082.i07}

Liversedge, M. (2020). GoldenCheetah OpenData
Project. OSF. \url{https://doi.org/10.17605/OSF.IO/6HFPZ}
}
