% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_management.R
\name{subset.gcod_db}
\alias{subset.gcod_db}
\title{Subset the local or remote perspective of a \code{gcod_db} object}
\usage{
\method{subset}{gcod_db}(x, subset, perspective = "remote")
}
\arguments{
\item{x}{an object of class \code{gcod_db}}

\item{subset}{logical expression indicating records to keep in either \code{remote(x)} or \code{local(x)} according to the value of \code{perspective};  missing values are taken as \code{FALSE}.}

\item{perspective}{either \code{"remote"} (default) or \code{"local"}, for the perspective to subset.}

\item{...}{currently not used.}
}
\description{
Subset the local or remote perspective of a \code{gcod_db} object
}
\details{
See \code{\link[=get_athlete_ids]{get_athlete_ids()}} for the variables each perspective holds.
}
\examples{
\donttest{
db <- get_athlete_ids()
## Return athletes from the remote perspective of db with "b7-9" in
## their IDs
subset(db, subset = grepl("b7-9", remote(db)$athlete_id), perspective = "remote")
}
}
